20-11-2024 club popup disabled done,create user realm db

+++ app/src/main/java/com/snofed/publicapp/models/UserResponse.kt
@@ -19,9 +19,9 @@ data class UserResponse(
 )
 
 data class userData(
-    val gender: Long,
-    val weight: Long,
-    val age: Long,
+    val gender: Int,
+    val weight: Int,
+    val age: Int,
     val isSubscribed: Boolean,
     val publicUserSettings: List<PublicUserSetting>,
     val favouriteClients: List<String>,
